requirements.txt
pillow==9.3.0
django-admin-interface==0.22.1
django-crispy-forms==1.14.0
django-allauth==0.51.0
django==4.1.3
django-autoslug==1.9.8
django-filter==22.1
paypal-checkout-serversdk==1.0.1